fix : 에러 수정
- Statistic 수익률 자료형 수정
- Exception import 패키지 수정
- OutpuView 숫자 천단위 콤마 추가
- GameController UI 로직 메소드 스태틱화

+++ src/main/java/lotto/domain/result/Statistic.java
@@ -11,7 +11,7 @@ public class Statistic {
     private final List<Lotto> userLotto;
     private final LuckyNumbers luckyNumbers;
     private final LinkedHashMap<Rank, Integer> result = new LinkedHashMap<>();
-    private Long profitRate;
+    private double profitRate;
 
     public Statistic(List<Lotto> userLotto, LuckyNumbers luckyNumbers,int amount){
         this.userLotto = userLotto;
@@ -39,14 +39,14 @@ public class Statistic {
             sum += (long) rank.getPrize() * result.get(rank);
         }
 
-        profitRate = sum / amount * 100;
+        profitRate = (double) sum / amount * 100;
     }
 
     public int getRankCount(Rank rank) {
         return result.getOrDefault(rank, 0);
     }
 
-    public long getProfitRate() {
+    public double getProfitRate() {
         return profitRate;
     }
